Crate pallet_evm_precompile_call_permit

Structs

CallPermitPrecompile

Precompile allowing to issue and dispatch call permits for gasless transactions. A user can sign a permit for a call that can be dispatched and paid by another user or smart contract.

Nonces

Storage prefix for nonces.

Enums

CallPermitPrecompileCall

Constants

CALL_DATA_LIMIT

PERMIT_TYPEHASH

EIP712 permit typehash.

Type Definitions

NoncesStorage

Storage type used to store EIP2612 nonces.